centralise classifier batch predict method
